Add full website to version control
It appears the only way to serve a website using GitHub Pages as a free
customer is via README.md or index.html in the root directory of the
repo.  Thus, I've added the built website to version control and a
simple redirect to that website.

Now that the website is fully generated via the build system, I've taken
the intermediate Markdown version of the tutorial off version control.